Added a simple food delivery app. Orders app has three methods (create, update, delete) that request authorization from the Auth app using the "get_auth" method, which requests authorization from the Auth app. The Auth app has a deliberate bug where authorization for deleting orders is refused, due to a typo.
Main considerations:

 Create, delete, and update have hard-coded order details. Do we want to change them to take
in parameters from the test function?

Right now, the Auth app returns a json object with many fields, include a "status" field. This is
meant to mimick an HTTP code. For example, any authorization requests made from the delete
method get a status of 424 (failed dependacy). In reality, the actual request returns a
code of 200, since delete successfully sent a request to the Auth app and received a response.